factor out Delta component
